Feature: Base class for data loading functions.
