+++ README.md
@@ -46,6 +46,6 @@ We need to divide the tasks among the team members in a way that will minimize h
 * 3/8/20 SG Branch calculator off dev, upload calculator project
 * 3/9/20 SG Add check for zero-length datafile to CsvReader
 * 3/9/20 SG Import coronavirus patient data for South Korea from https://www.kaggle.com/kimjihoo/coronavirusdataset#patient.csv
-
+* 3/11/20 SG Separate data loader functionality into its own base class
